Fixes issue #266 (#267)
- ingress gateway service wasn't properly reconciled with NodePort type
