+++ pkg/controller/remoteistio/remoteistio_controller.go
@@ -289,7 +289,7 @@ func (r *ReconcileRemoteConfig) reconcile(remoteConfig *istiov1beta1.RemoteIstio
 	err = cluster.Reconcile(remoteConfig, istio)
 	if err != nil {
 		err = emperror.Wrap(err, "could not reconcile remote istio")
-		if _, ok := errors.Cause(err).(remoteclusters.IngressSetupPendingError); ok {
+		if _, ok := errors.Cause(err).(k8sutil.IngressSetupPendingError); ok {
 			updateRemoteConfigStatus(r.Client, remoteConfig, istiov1beta1.ReconcileFailed, errors.Cause(err).Error(), logger)
 			return reconcile.Result{
 				Requeue:      true,
